refactor: move inputs to a common folder
