+++ frontend/src/features/common/inputs/InputSlider.tsx
